[TS-276] Rename: getOutlineText.style 파일 이름을 Star.style로 변경
현재 파일 이름 컨벤션에 맞춰 변경함
