+++ src/components/StarPage/StarMain/Star.style.ts
